Collect the configuration line editors in system::conf
patch_conf_array, set_conf_value and set_conf_line lived in the
mkinitcpio module although bootmenu.rs also edits ZFSBootMenu's copy of
that file, and set_option and set_parallel_downloads did the same job for
pacman.conf from the repository module. One module holds them all, with
their tests; behaviour is unchanged, including set_parallel_downloads
appending a missing key at the end rather than under [options].
